pyproject.toml
[project]
name = "hjrs-thesis-karen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anywidget>=0.9.18",
    "datetime>=5.5",
    "dotenv>=0.9.9",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "hdbscan>=0.8.40",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "joblib>=1.5.1",
    "matplotlib>=3.10.3",
    "msoffcrypto-tool>=5.4.2",
    "nbformat>=5.10.4",
    "openai>=1.101.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pingouin>=0.5.5",
    "plotly>=6.2.0",
    "rapidfuzz>=3.13.0",
    "scikit-learn==1.5.2",
    "scikit-posthocs>=0.11.4",
    "seaborn>=0.13.2",
    "selenium>=4.35.0",
    "sentence-transformers>=5.0.0",
    "statsmodels>=0.14.5",
    "streamlit>=1.47.1",
    "torch>=2.7.1",
    "umap-learn>=0.5.7",
    "webdriver-manager>=4.0.2",
]
